requirement.txt
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.6
django-environ>=0.10.0
mysqlclient>=2.2.0
celery>=5.3.0

pip freeze > requirements.txt


